+++ pg_documentdb/include/metadata/metadata_cache.h
@@ -183,7 +183,6 @@ Oid PostgresMakeIntervalFunctionId(void);
 Oid PostgresDateBinFunctionId(void);
 Oid PostgresAgeBetweenTimestamp(void);
 Oid PostgresDatePartFromInterval(void);
-Oid PostgresUUIDInFunctionId(void);
 
 /* Index AM */
 Oid RumIndexAmId(void);