+++ pg_documentdb/src/metadata/metadata_cache.c
@@ -564,9 +564,6 @@ typedef struct DocumentDBApiOidCacheData
 	/* Oid of the extract_interval postgres function which extracts a given date part from interval. */
 	Oid PostgresDatePartFromInterval;
 
-	/* OID of the uuid_in postgres method which converts a string to uuid */
-	Oid PostgresUUIDInFunctionIdOid;
-
 	/* OID of Rum Index access methods */
 	Oid RumIndexAmId;
 
@@ -2643,16 +2640,6 @@ PostgresDatePartFromInterval(void)
 {
 	return GetPostgresInternalFunctionId(&Cache.PostgresDatePartFromInterval,
 										 "interval_part");
-}
-
-
-/*
- * Returns the OID of the "uuid_in" internal postgres method
- */
-Oid
-PostgresUUIDInFunctionId(void)
-{
-	return GetPostgresInternalFunctionId(&Cache.PostgresUUIDInFunctionIdOid, "uuid_in");
 }