+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### documentdb v0.103-0 (Unreleased) ###
+* Support `$convert` on `binData` to `binData`, `string` to `binData` and `binData` to `string` (except with `format: auto`) *[Feature]*
 
 ### documentdb v0.102-0 (March 26, 2025) ###
 * Support index pushdown for vector search queries *[Bugfix]*